python/requirements.txt
streamlit
numpy
plotly
pandas
yfinance
mysql-connector-python
python-dateutil